Relocate directory-sorting behavior to Packs.for_file
The RSpec integration in the `packs-rails` gem depends on `Packs.all` to
return a collection of directory names representing packages.

https://github.com/rubyatscale/packs-rails/blob/505fceb9ff50f666ee8c9c0ffabf1037c2519c43/lib/packs/rails/integrations/rspec.rb#L18

The current implementation of `Packs.all` returns a collection of
package names sorted by the length of their directory entries, a
behavior that `Packs.for_file` relies on. However, this
out-of-lexical-order collection creates problems for RSpec in that fails
to run any spec files found in directories not listed in the expected
alphabetical order. [This Github issue goes into
moredetail.](https://github.com/rubyatscale/packs-rails/issues/58)

The solution to this problem looks to be straightforward: relocate the
block that sorts package names by their length to the implementation of
`Packs.for_file`, leaving `Packs.all` to return package names in the
same order they were read off the filesystem.